requirements.txt
gadlint
gadignore
gads3
gadutils
gadasyncblock
gadlogging
gadsqlalchemy
gadfastcron
gadfastrouter
gadfastsentry
gadfastetcd
gadfasthealth
gadfastopenapi
gadpytestprofiler
gadcodegenerator
uvicorn
pydantic
pydantic-settings
asyncpg
psycopg2
protobuf==3.20
pytest-asyncio
alembic